+++ src/main/java/mod/gottsch/forge/gottschcore/block/entity/StructureMobSetBlockEntity.java
@@ -19,11 +19,16 @@ import mod.gottsch.forge.gottschcore.GottschCore;
 import mod.gottsch.forge.gottschcore.util.ModUtil;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.ListTag;
+import net.minecraft.nbt.StringTag;
+import net.minecraft.nbt.Tag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 /**
@@ -38,6 +43,7 @@ public abstract class StructureMobSetBlockEntity extends BlockEntity {
     public static final String TARGET_BLOCK = "targetBlock";
 
     private ResourceLocation mobSet;
+    private List<ResourceLocation> mobSets;
     private int proximity;
     private ResourceLocation targetBlock;
 
@@ -54,7 +60,14 @@ public abstract class StructureMobSetBlockEntity extends BlockEntity {
                 if (tag.contains(MOBSET)) {
                     this.mobSet = ModUtil.asLocation(tag.getString(MOBSET)).orElse(null);
                 }
-                // TODO load MOBSETS
+                if (tag.contains(MOBSETS, Tag.TAG_LIST)) {
+                    ListTag listTag = tag.getList(MOBSETS, Tag.TAG_STRING);
+                    listTag.stream()
+                            .map(Tag::getAsString)
+                            .filter(str -> !str.isEmpty())
+                            .map(ModUtil::asLocation)
+                            .forEach(location -> location.ifPresent(this.getMobSets()::add));
+                }
 
                 if (tag.contains(PROXIMITY)) {
                     this.proximity = tag.getInt(PROXIMITY);
@@ -82,6 +95,12 @@ public abstract class StructureMobSetBlockEntity extends BlockEntity {
             Optional.ofNullable(getMobSet())
                     .ifPresent(mobSet -> nbt.putString(MOBSET, getMobSet().toString()));
 
+            if (getMobSets() != null && !getMobSets().isEmpty()) {
+                ListTag list = new ListTag();
+                getMobSets().forEach(name -> list.add(StringTag.valueOf(name.toString())));
+                nbt.put(MOBSETS, list);
+            }
+
             nbt.putInt(PROXIMITY, this.proximity);
 
             Optional.ofNullable(getTargetBlock())
@@ -99,6 +118,14 @@ public abstract class StructureMobSetBlockEntity extends BlockEntity {
 
     public void setMobSet(ResourceLocation mobSet) {
         this.mobSet = mobSet;
+    }
+
+    public List<ResourceLocation> getMobSets() {
+        return mobSets != null ? mobSets : (mobSets = new ArrayList<>());
+    }
+
+    public void setMobSets(List<ResourceLocation> mobSets) {
+        this.mobSets = mobSets;
     }
 
     public int getProximity() {